% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{prepdata}
\alias{prepdata}
\title{Reformat data to prepare it for writing functions}
\usage{
prepdata(x, lociprefix)
}
\arguments{
\item{x}{a tibble imported by read_GenAlEx().}

\item{lociprefix}{a character string specifying the first letters shared by all loci in the data used to identify locus columns.}
}
\value{
a tibble
}
\description{
What the title said
}
