% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_GenAlEx.R
\name{read_GenAlEx}
\alias{read_GenAlEx}
\title{Import from GenAlEx}
\usage{
read_GenAlEx(path)
}
\arguments{
\item{path}{a character string specifying the path to the file that should be imported.}
}
\value{
a tibble
}
\description{
Reads the subset of the fields in a GenAlEx-formatted .xlsx-file that are relevant for the package.
}
