% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_locus_file.R
\name{write_locus_file}
\alias{write_locus_file}
\title{Write locus file}
\usage{
write_locus_file(x, lociprefix, locus_filename)
}
\arguments{
\item{x}{A tibble imported by msatwRangler::read_GenAlEx().}

\item{lociprefix}{A character string specifying the first letters shared by all loci in the data used to identify locus columns.}

\item{locus_filename}{The desired filename of the output file.}
}
\value{
A .txt file placed in the directory
}
\description{
Create a .txt file that meets the input requirements for the locus file used in MicroMerger.
}
