% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pedigree_file.R
\name{write_pedigree_file}
\alias{write_pedigree_file}
\title{Write pedigree file}
\usage{
write_pedigree_file(x, pedigree_filename)
}
\arguments{
\item{x}{a tibble formatted by msatwRangler::prepdata().}

\item{pedigree_filename}{the desired filename of the output file}
}
\value{
a .txt file placed in the directory.
}
\description{
Create a .txt file that meets the input file requirements of MicroMerger.
}
