% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_GenAlEx.R
\name{write_GenAlEx}
\alias{write_GenAlEx}
\title{Write a GenAlEx formated .csv file}
\usage{
write_GenAlEx(path)
}
\arguments{
\item{path}{The desired path for the output file.}
}
\value{
a .csv file placed in the directory
}
\description{
Create a GenAlEx formated .csv file in the directory from a dataset imported from Geneious.
}
